+++ dev/tests/integration/testsuite/Magento/ConfigurableProduct/_files/customer_quote_with_items_configurable_product_rollback.php
@@ -7,10 +7,9 @@ declare(strict_types=1);
 
 use Magento\Framework\Registry;
 use Magento\Quote\Api\CartRepositoryInterface;
-use Magento\Quote\Model\ResourceModel\Quote\Collection;
-use Magento\Quote\Model\ResourceModel\Quote\CollectionFactory;
 use Magento\TestFramework\Helper\Bootstrap;
 use Magento\TestFramework\ObjectManager;
+use Magento\TestFramework\Quote\Model\GetQuoteByReservedOrderId;
 
 /** @var $objectManager ObjectManager */
 $objectManager = Bootstrap::getObjectManager();
@@ -20,12 +19,10 @@ $registry = $objectManager->get(Registry::class);
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
-/** @var CartRepositoryInterface $quoteRepository */
-$quoteRepository = $objectManager->get(CartRepositoryInterface::class);
-/** @var Collection $quoteCollection */
-$quoteCollection = $objectManager->create(CollectionFactory::class)->create();
-$quoteCollection->addFieldToFilter('reserved_order_id', 'customer_quote_configurable_products');
-foreach ($quoteCollection as $quote) {
+$quote = $objectManager->get(GetQuoteByReservedOrderId::class)->execute('customer_quote_configurable_products');
+if ($quote !== null) {
+    /** @var CartRepositoryInterface $quoteRepository */
+    $quoteRepository = $objectManager->get(CartRepositoryInterface::class);
     $quoteRepository->delete($quote);
 }
